refctor code
