+++ classroom.rb
@@ -10,5 +10,5 @@ class Classroom
   def add_student(student)
     @students.push(student)
     student.classroom = self
-  end 
+  end
 end
